+++ CogniteSdk.Types/Relationships/RelationshipFilter.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using CogniteSdk.Types.Common;
 
-namespace CogniteSdk.Beta
+namespace CogniteSdk
 {
     /// <summary>
     /// Class for filtering relationships.
