+++ CogniteSdk/src/Resources/Beta.cs
@@ -17,7 +17,6 @@ namespace CogniteSdk.Resources
         /// <summary>
         /// Client Relationships extension methods
         /// </summary>
-        public Beta.RelationshipResource Relationships  { get; }
 
         /// <summary>
         /// Will only be instantiated by the client.
@@ -26,7 +25,6 @@ namespace CogniteSdk.Resources
         /// <param name="ctx">Context to use for the request.</param>
         internal BetaResource(Func<CancellationToken, Task<string>> authHandler, HttpContext ctx) : base(authHandler, ctx)
         {
-            Relationships = new Beta.RelationshipResource(authHandler, ctx);
         }
     }
 }
