+++ CogniteSdk/src/Resources/Relationships.cs
@@ -6,12 +6,12 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-using static Oryx.Cognite.Beta.HandlerModule;
+using static Oryx.Cognite.HandlerModule;
 using HttpContext = Oryx.Context<Microsoft.FSharp.Core.Unit>;
 
-namespace CogniteSdk.Resources.Beta
+namespace CogniteSdk.Resources
 {
-    using CogniteSdk.Beta;
+    using CogniteSdk;
 
     /// <summary>
     /// For internal use. Contains all relationship methods.
@@ -40,7 +40,7 @@ namespace CogniteSdk.Resources.Beta
                 throw new ArgumentNullException(nameof(query));
             }
 
-            var req = Oryx.Cognite.Beta.Relationships.list<ItemsWithCursor<Relationship>>(query);
+            var req = Oryx.Cognite.Relationships.list<ItemsWithCursor<Relationship>>(query);
             return await RunAsync(req, token).ConfigureAwait(false);
         }
 
@@ -57,7 +57,7 @@ namespace CogniteSdk.Resources.Beta
                 throw new ArgumentNullException(nameof(relationships));
             }
 
-            var req = Oryx.Cognite.Beta.Relationships.create<IEnumerable<Relationship>>(relationships);
+            var req = Oryx.Cognite.Relationships.create<IEnumerable<Relationship>>(relationships);
             return await RunAsync(req, token).ConfigureAwait(false);
         }
 
@@ -74,7 +74,7 @@ namespace CogniteSdk.Resources.Beta
                 throw new ArgumentNullException(nameof(externalIds));
             }
 
-            var req = Oryx.Cognite.Beta.Relationships.delete<EmptyResponse>(externalIds);
+            var req = Oryx.Cognite.Relationships.delete<EmptyResponse>(externalIds);
             return await RunAsync(req, token).ConfigureAwait(false);
         }
 
@@ -91,7 +91,7 @@ namespace CogniteSdk.Resources.Beta
                 throw new ArgumentNullException(nameof(ids));
             }
 
-            var req = Oryx.Cognite.Beta.Relationships.retrieve<IEnumerable<Relationship>>(ids);
+            var req = Oryx.Cognite.Relationships.retrieve<IEnumerable<Relationship>>(ids);
             return await RunAsync(req, token).ConfigureAwait(false);
         }
     }
