+++ Oryx.Cognite/src/Oryx.Cognite.fsproj
@@ -19,13 +19,13 @@
     <Compile Include="DataPoints.fs" />
     <Compile Include="Events.fs" />
     <Compile Include="Raw.fs" />
+    <Compile Include="Relationships.fs" />
     <Compile Include="Files.fs" />
     <Compile Include="Sequences.fs" />
     <Compile Include="Login.fs" />
     <Compile Include="Playground/Handler.fs" />
     <Compile Include="Playground/Functions.fs" />
     <Compile Include="Beta/Handler.fs" />
-    <Compile Include="Beta/Relationships.fs" />
     <Compile Include="Beta\Domains.fs" />
     <None Include="..\..\LICENSE" Pack="true" Visible="false" PackagePath="" />
   </ItemGroup>
